# path: make Base inlineable