--- a/src/cmd/compile/internal/test/inl_test.go
+++ b/src/cmd/compile/internal/test/inl_test.go
@@ -234,6 +234,7 @@ func TestIntendedInlining(t *testing.T) {
 			"(*B).Loop",
 		},
 		"path": {
+			"Base",
 			"scanChunk",
 		},
 		"path/filepath": {